requirements.txt
# Smart Code Reviewer - Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.5.3
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.26.0

